initial workings to swap page location in flow
