+++ src/ecrc-frontend/src/components/page/applicationForm/ApplicationForm.js
@@ -101,7 +101,8 @@ export default function ApplicationForm({
   const [provinces, setProvinces] = useState([]);
 
   useEffect(() => {
-    if (!isAuthorized() || !isActionPerformed("consent")) setToHome(true);
+    if (!isAuthorized() || !isActionPerformed("userConfirmation"))
+      setToHome(true);
 
     const token = sessionStorage.getItem("jwt");
     const uuid = sessionStorage.getItem("uuid");
