+++ src/ecrc-frontend/src/components/page/consent/Consent.js
@@ -1,5 +1,6 @@
 /* eslint-disable react/jsx-one-expression-per-line */
 import React, { useState, useEffect } from "react";
+import axios from "axios";
 import { Redirect } from "react-router-dom";
 import PropTypes from "prop-types";
 
@@ -16,10 +17,51 @@ import {
   isActionPerformed
 } from "../../../modules/AuthenticationHelper";
 
-export default function Consent({ page: { header, setError } }) {
+export default function Consent({
+  page: {
+    header,
+    applicant: {
+      legalFirstNm,
+      legalSecondNm,
+      legalSurnameNm,
+      alias1FirstNm,
+      alias1SecondNm,
+      alias1SurnameNm,
+      alias2FirstNm,
+      alias2SecondNm,
+      alias2SurnameNm,
+      alias3FirstNm,
+      alias3SecondNm,
+      alias3SurnameNm,
+      birthDt,
+      genderTxt,
+      countryNm,
+      mailingAddressLine1,
+      mailingCity,
+      mailingProvince,
+      mailingPostalCode,
+      birthPlace,
+      driversLicNo,
+      phoneNumber,
+      jobTitle
+    },
+    org: {
+      orgApplicantRelationship,
+      orgTicketNumber,
+      defaultScheduleTypeCd,
+      defaultCrcScopeLevelCd
+    },
+    setApplicationInfo,
+    saveApplicant,
+    saveOrg,
+    saveApplicationInfo,
+    setError
+  }
+}) {
   const [toAppHome, setToAppHome] = useState(false);
   const [toHome, setToHome] = useState(false);
-  const [toApplicationForm, setToApplicationForm] = useState(false);
+  const [loading, setLoading] = useState(false);
+  const [toSuccess, setToSuccess] = useState(false);
   const [toError, setToError] = useState(false);
   const [firstBoxChecked, setFirstBoxChecked] = useState(false);
   const [secondBoxChecked, setSecondBoxChecked] = useState(false);
@@ -29,7 +71,7 @@ export default function Consent({ page: { header, setError } }) {
   useEffect(() => {
     window.scrollTo(0, 0);
 
-    if (!isAuthorized() || !isActionPerformed("userConfirmation")) {
+    if (!isAuthorized() || !isActionPerformed("infoReview")) {
       setToAppHome(true);
     }
   }, []);
@@ -54,7 +96,169 @@ export default function Consent({ page: { header, setError } }) {
     buttonStyle: "btn ecrc_go_btn",
     buttonSize: "btn",
     type: "submit",
-    disabled: !continueBtnEnabled
+    disabled: !continueBtnEnabled,
+    loader: loading
+  };
+
+  const confirm = () => {
+    setLoading(true);
+
+    if (!isAuthorized()) {
+      setError("session expired");
+      setToError(true);
+      return;
+    }
+    const token = sessionStorage.getItem("jwt");
+    const uuid = sessionStorage.getItem("uuid");
+
+    const createApplicantInfo = {
+      orgTicketNumber,
+      requestGuid: uuid,
+      callPurpose: "CRC",
+      legalSurnameNm,
+      legalFirstNm,
+      legalSecondNm,
+      birthDt,
+      genderTxt,
+      birthPlace,
+      alias1FirstNm,
+      alias1SecondNm,
+      alias1SurnameNm,
+      alias2FirstNm,
+      alias2SecondNm,
+      alias2SurnameNm,
+      alias3FirstNm,
+      alias3SecondNm,
+      alias3SurnameNm,
+      phoneNumber,
+      addressLine1: mailingAddressLine1,
+      cityNm: mailingCity,
+      provinceNm: mailingProvince,
+      countryNm,
+      postalCodeTxt: mailingPostalCode,
+      driversLicNo
+    };
+
+    let partyId;
+    let sessionId;
+    let invoiceId;
+    let serviceFeeAmount;
+    let serviceId;
+
+    Promise.all([
+      axios.post("/ecrc/private/createApplicant", createApplicantInfo, {
+        headers: {
+          Authorization: `Bearer ${token}`
+        }
+      }),
+      axios.get(
+        `/ecrc/private/getNextSessionId?orgTicketNumber=${orgTicketNumber}&requestGuid=${uuid}`,
+        {
+          headers: {
+            Authorization: `Bearer ${token}`
+          }
+        }
+      ),
+      axios.get(
+        `/ecrc/private/getNextInvoiceId?orgTicketNumber=${orgTicketNumber}&requestGuid=${uuid}`,
+        {
+          headers: {
+            Authorization: `Bearer ${token}`
+          }
+        }
+      ),
+      axios.get(
+        `/ecrc/private/getServiceFeeAmount?orgTicketNumber=${orgTicketNumber}&scheduleTypeCd=${defaultScheduleTypeCd}&scopeLevelCd=${defaultCrcScopeLevelCd}&requestGuid=${uuid}`,
+        {
+          headers: {
+            Authorization: `Bearer ${token}`
+          }
+        }
+      )
+    ])
+      .then(all => {
+        partyId = all[0].data.partyId;
+        sessionId = all[1].data.sessionId;
+        invoiceId = all[2].data.invoiceId;
+        serviceFeeAmount = all[3].data.serviceFeeAmount;
+
+        // NEED CLARIFICATION: - as per Jason Lee, awaiting confirmation
+        // eivPassDetailsResults - String returned from equifax, see Shaun
+        const newCRC = {
+          orgTicketNumber,
+          requestGuid: uuid,
+          schedule_Type_Cd: defaultScheduleTypeCd,
+          scope_Level_Cd: defaultCrcScopeLevelCd,
+          appl_Party_Id: partyId,
+          org_Appl_To_Pay: "A",
+          applicant_Posn: jobTitle,
+          child_Care_Fac_Nm: "child_Care_Fac_Nm",
+          governing_Body_Nm: "governing_Body_Nm",
+          session_Id: sessionId,
+          invoice_Id: invoiceId,
+          auth_Release_EIV_Vendor_YN: "Y",
+          auth_Conduct_CRC_Check_YN: "Y",
+          auth_Release_To_Org_YN: "Y",
+          appl_Identity_Verified_EIV_YN: "Y",
+          eivPassDetailsResults: "eivPassDetailsResults"
+        };
+
+        return axios.post("/ecrc/private/createNewCRCService", newCRC, {
+          headers: {
+            Authorization: `Bearer ${token}`
+          }
+        });
+      })
+      .then(crcResponse => {
+        serviceId = crcResponse.data.serviceId;
+
+        const appInfo = {
+          partyId,
+          sessionId,
+          invoiceId,
+          serviceFeeAmount,
+          serviceId
+        };
+
+        setApplicationInfo(appInfo);
+
+        if (orgApplicantRelationship === "VOLUNTEER") {
+          setToSuccess(true);
+        } else {
+          saveApplicant();
+          saveOrg();
+          saveApplicationInfo(appInfo);
+
+          const createURL = {
+            invoiceNumber: invoiceId,
+            requestGuid: uuid,
+            approvedPage: `${process.env.REACT_APP_FRONTEND_BASE_URL}/ecrc/success`,
+            declinedPage: `${process.env.REACT_APP_FRONTEND_BASE_URL}/ecrc/success`,
+            errorPage: `${process.env.REACT_APP_FRONTEND_BASE_URL}/ecrc/success`,
+            totalItemsAmount: serviceFeeAmount,
+            serviceIdRef1: serviceId,
+            partyIdRef2: partyId
+          };
+
+          const currentPayload = accessJWTToken(sessionStorage.getItem("jwt"));
+          const newPayload = {
+            ...currentPayload,
+            actionsPerformed: [...currentPayload.actionsPerformed, "infoReview"]
+          };
+          generateJWTToken(newPayload);
+
+          axios
+            .post("/ecrc/private/createPaymentUrl", createURL, {
+              headers: {
+                Authorization: `Bearer ${token}`
+              }
+            })
+            .then(urlResponse => {
+              window.location.href = urlResponse.data.paymentUrl;
+            });
+        }
+      })
+      .catch(() => {});
   };
 
   if (toHome) {
@@ -65,20 +269,19 @@ export default function Consent({ page: { header, setError } }) {
     return <Redirect to="/" />;
   }
 
-  if (toApplicationForm) {
+  if (toSuccess) {
     if (!isAuthorized()) {
       setError("session expired");
       return setToError(true);
     }
 
     const currentPayload = accessJWTToken(sessionStorage.getItem("jwt"));
-    const actionsPerformed = [...currentPayload.actionsPerformed, "consent"];
     const newPayload = {
       ...currentPayload,
-      actionsPerformed
+      actionsPerformed: [...currentPayload.actionsPerformed, "consent"]
     };
     generateJWTToken(newPayload);
-    return <Redirect to="/criminalrecordcheck/applicationform" />;
+    return <Redirect to="/criminalrecordcheck/success" />;
   }
 
   const asterisk = (
@@ -116,12 +319,7 @@ export default function Consent({ page: { header, setError } }) {
                 setToHome(true);
               }}
             />
-            <Button
-              button={continueButton}
-              onClick={() => {
-                setToApplicationForm(true);
-              }}
-            />
+            <Button button={continueButton} onClick={confirm} />
           </div>
         </div>
