+++ src/ecrc-frontend/src/components/page/consent/Consent.test.js
@@ -19,10 +19,50 @@ describe("Consent Page Component", () => {
     name: "Criminal Record Check"
   };
 
+  const setApplicationInfo = jest.fn();
+  const saveApplicant = jest.fn();
+  const saveOrg = jest.fn();
+  const saveApplicationInfo = jest.fn();
   const setError = jest.fn();
+
+  const applicant = {
+    legalFirstNm: "Robert",
+    legalSecondNm: "Norman",
+    legalSurnameNm: "Ross",
+    birthPlace: "Daytona Beach, Florida",
+    birthDt: "1942-10-29",
+    genderTxt: "Male",
+    driversLicNo: "1234567",
+    phoneNumber: "2501234567",
+    emailAddress: "bob.ross@example.com",
+    addressLine1: "123 Somewhere",
+    cityNm: "Here",
+    provinceNm: "British Columbia",
+    postalCodeTxt: "V9V 9V9",
+    countryNm: "Canada",
+    mailingAddressLine1: "456 Elsewhere",
+    mailingCity: "There",
+    mailingProvince: "Ontario",
+    mailingPostalCode: "V1V 1A1",
+    jobTitle: "Painter",
+    organizationFacility: ""
+  };
+
+  const org = {
+    orgApplicantRelationship: "EMPLOYEE",
+    orgTicketNumber: "crce",
+    defaultScheduleTypeCd: "WBSD",
+    defaultCrcScopeLevelCd: "WWCH"
+  };
 
   const page = {
     header,
+    applicant,
+    org,
+    setApplicationInfo,
+    saveApplicant,
+    saveOrg,
+    saveApplicationInfo,
     setError
   };
